+++ lib/td_auth/permissions/permission_group.ex
@@ -7,13 +7,28 @@ defmodule TdAuth.Permissions.PermissionGroup do
 
   import Ecto.Changeset
 
+  alias TdAuth.Permissions.Constants
   alias TdAuth.Permissions.Permission
+
+  @custom_prefix Constants.custom_prefix()
 
   schema "permission_groups" do
     field(:name, :string)
     has_many(:permissions, Permission, on_replace: :nilify)
 
     timestamps()
+  end
+
+  def changeset_external(%__MODULE__{} = permission_group, params) do
+    permission_group
+    |> changeset(params)
+    |> validate_name()
+  end
+
+  def changeset_external(params) do
+    params
+    |> changeset()
+    |> validate_name()
   end
 
   def changeset(params) do
@@ -35,4 +50,18 @@ defmodule TdAuth.Permissions.PermissionGroup do
       message: "group.delete.existing.permissions"
     )
   end
+
+  defp validate_name(%{changes: %{name: name}} = changeset) do
+    if String.starts_with?(name, @custom_prefix) do
+      changeset
+    else
+      add_error(
+        changeset,
+        :name,
+        "External permission group creation requires a name starting with '#{@custom_prefix}'"
+      )
+    end
+  end
+
+  defp validate_name(changeset), do: changeset
 end