+++ lib/td_auth_web/controllers/permission_controller.ex
@@ -43,8 +43,100 @@ defmodule TdAuthWeb.PermissionController do
     claims = conn.assigns[:current_resource]
 
     with %Permission{} = permission <- Permissions.get_permission!(id),
-         {:can, true} <- {:can, can?(claims, view(permission))} do
+         {:can, true} <- {:can, can?(claims, view(Permission))} do
       render(conn, "show.json", permission: permission)
+    end
+  end
+
+  swagger_path :create do
+    description("Creates a new permission")
+    produces("application/json")
+
+    parameters do
+      configuration(
+        :body,
+        Schema.ref(:CreatePermission),
+        "Parameters used to create a permission"
+      )
+    end
+
+    response(200, "OK", Schema.ref(:PermissionResponse))
+    response(403, "Forbidden")
+    response(422, "Client Error")
+  end
+
+  def create(
+        conn,
+        %{
+          "permission" => %{
+            "name" => _name,
+            "permission_group_id" => _permission_group_id
+          },
+          "allow_non_custom_name" => true
+        } = params
+      ) do
+    create(
+      conn,
+      params,
+      &Permissions.create_permission/1
+    )
+  end
+
+  def create(
+        conn,
+        %{
+          "permission" => %{
+            "name" => _name,
+            "permission_group_id" => _permission_group_id
+          }
+        } = params
+      ) do
+    create(
+      conn,
+      params,
+      &Permissions.create_external_permission/1
+    )
+  end
+
+  defp create(
+         conn,
+         %{
+           "permission" =>
+             %{"name" => _name, "permission_group_id" => _permission_group_id} = permission_params
+         },
+         fn_create_permission
+       ) do
+    with claims <- conn.assigns[:current_resource],
+         {:can, true} <- {:can, can?(claims, create(Permission))},
+         {:ok, %Permission{} = permission} <-
+           fn_create_permission.(permission_params) do
+      conn
+      |> put_status(:created)
+      |> put_resp_header("location", Routes.permission_path(conn, :show, permission))
+      |> render("show.json", permission: permission)
+    end
+  end
+
+  swagger_path :delete do
+    description("Delete Permission")
+    produces("application/json")
+
+    parameters do
+      id(:path, :integer, "Permission ID", required: true)
+    end
+
+    response(204, "No Content")
+    response(403, "Forbidden")
+    response(422, "Unprocessable Entity")
+  end
+
+  def delete(conn, %{"id" => id}) do
+    claims = conn.assigns[:current_resource]
+
+    with %Permission{} = permission <- Permissions.get_permission!(id),
+         {:can, true} <- {:can, can?(claims, delete(Permission))},
+         {:ok, _deleted_permission} <- Permissions.delete_permission(permission) do
+      send_resp(conn, :no_content, "")
     end
   end
 end