+++ priv/repo/migrations/20220818144535_add_id_timestamps_to_roles_permissions.exs
@@ -1,0 +1,10 @@
+defmodule TdAuth.Repo.Migrations.AddIdTimestampsToRolesPermissions do
+  use Ecto.Migration
+
+  def change do
+    alter table(:roles_permissions) do
+      add(:id, :bigserial, primary_key: true)
+      timestamps(default: "#{DateTime.utc_now()}")
+    end
+  end
+end